[1-12-2020] change method remove for all provider, change structure schema database

+++ src/server/core/File.php
@@ -13,13 +13,15 @@ class File {
    * @param string $filename Filename for moving
    */
   public static function moveUploaded(array $uploadedFile, string $filename) {
-    if (Request::getInstance()->getMethod() == 'post') {
-      if (!move_uploaded_file($uploadedFile["tmp_name"], __ROOT__ . $filename)) {
-        throw new Exception("Error when moving uploaded file");
-      }
-    } else {
-      if (!rename($uploadedFile["tmp_name"], __ROOT__ . $filename)) {
-        throw new Exception("Error when moving uploaded file");
+    if (file_exists($uploadedFile["tmp_name"])) {
+      if (Request::getInstance()->getMethod() == 'post') {
+        if (!move_uploaded_file($uploadedFile["tmp_name"], __ROOT__ . $filename)) {
+          throw new Exception("Error when moving uploaded file");
+        }
+      } else {
+        if (!rename($uploadedFile["tmp_name"], __ROOT__ . $filename)) {
+          throw new Exception("Error when moving uploaded file");
+        }
       }
     }
   }
@@ -42,8 +44,10 @@ class File {
    * @param string $fileName Filename for deleting
    */
   public static function delete(string $fileName) {
-    if (!unlink(__ROOT__ . $fileName)) {
-      throw new Exception("Error when deleting file");
+    if (file_exists($fileName)) {
+      if (!unlink(__ROOT__ . $fileName)) {
+        throw new Exception("Error when deleting file");
+      }
     }
   }
 }
